# docs/specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: chatlab HTTP API
  version: 1.0.0-rc.1
  description: |
    chatlab is a local development platform for chat agents — workspaces, multi-provider LLMs,
    feedback corpus.

    All endpoints under `/v1/*` require `Authorization: Bearer <token>`. Token is permissive by
    default; set `CHATLAB_REQUIRE_TOKEN=<value>` for strict mode. `/healthz`, `/readyz`,
    `/favicon.ico` are unauthenticated.

    Errors are returned as `{ error: { message, type, code, error_subcode?, error_data?, request_id } }`.
  license:
    name: MIT
    url: https://github.com/jvrmaia/chatlab/blob/main/LICENSE
  contact:
    url: https://github.com/jvrmaia/chatlab
servers:
  - url: http://127.0.0.1:4480
    description: Default local chatlab
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
  parameters:
    WorkspaceId:
      in: path
      name: id
      required: true
      schema: { type: string, format: uuid }
      description: Workspace UUID.
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Conflict with current resource state
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Workspace:
      type: object
      required: [id, nickname, storage_type, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        nickname: { type: string }
        storage_type:
          type: string
          enum: [memory, sqlite, duckdb]
        storage_path:
          type: string
          description: "Absolute path to the workspace's data file. Absent for `memory`; present for `sqlite` / `duckdb`."
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    WorkspaceCreate:
      type: object
      required: [nickname, storage_type]
      properties:
        nickname: { type: string }
        storage_type:
          type: string
          enum: [memory, sqlite, duckdb]
    WorkspaceDeleted:
      type: object
      required: [removed_id, active]
      properties:
        removed_id: { type: string, format: uuid }
        active: { $ref: "#/components/schemas/Workspace" }
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code, request_id]
          properties:
            message: { type: string }
            type: { type: string, description: "Always `ApiError`." }
            code: { type: integer }
            error_subcode:
              oneOf:
                - { type: string }
                - { type: integer }
            error_data:
              type: object
              additionalProperties: true
            request_id: { type: string }
security:
  - bearer: []
paths:
...
  /v1/workspaces:
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      summary: List workspaces + active id
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [data, active_id]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Workspace" }
                  active_id: { type: string, format: uuid }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      summary: Create a workspace
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceCreate" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/workspaces/active:
    get:
      tags: [Workspaces]
      operationId: getActiveWorkspace
      summary: Read the currently-active workspace
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /v1/workspaces/{id}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Workspaces]
      operationId: getWorkspace
      summary: Read a workspace
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Workspaces]
      operationId: patchWorkspace
      summary: Rename a workspace (`storage_type` and `storage_path` are immutable)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nickname: { type: string }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Workspaces]
      operationId: deleteWorkspace
      summary: Delete a workspace + its data files
      parameters:
        - in: query
          name: confirm
          required: true
          schema: { type: string, enum: ["true"] }
          description: Required typed confirmation.
      responses:
        "200":
          description: deleted; the new active workspace is returned
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceDeleted" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /v1/workspaces/{id}/activate:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      tags: [Workspaces]
      operationId: activateWorkspace
      summary: Switch the running adapter to this workspace (idempotent)
      description: |
        If the AgentRunner has in-flight calls, the activation waits up to 2 s for them to drain.
        If they don't, the activation fails with `409 ZZ_WORKSPACE_BUSY` and the client should retry.
      responses:
        "200":
          description: activated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...